+++ CHANGELOG.rst
@@ -2,8 +2,8 @@
 Changelog
 =========
 
-v0.8.dev
---------
+v0.8 (2020-10-06)
+-----------------
 
 * Remove deprecated ignore_cache parameter.
 * Fix query execution ID not cached locally when cached remotely.